table/tables: improve package unit test code coverage to 82% (#10366)
